requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.14.0
bleach==6.2.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
llvmlite==0.45.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mistune==3.1.4
more-itertools==10.8.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
numba==0.62.0
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.109.1
openai-whisper==20250625
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyttsx3==2.99
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tiktoken==0.11.0
tinycss2==1.4.0
torch==2.8.0
torchaudio==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
yarg==0.1.9
